+++ ricaun.AutoCAD.UI/Utils/RibbonThemePanelUtils.cs
@@ -105,6 +105,11 @@ namespace ricaun.AutoCAD.UI.Utils
                         foreach (var item in ribbonRowPanel.Items)
                             UpdateImageThemes(item, isLight);
                         break;
+                    case RibbonList ribbonList:
+                        UpdateRibbonItemTheme(ribbonList);
+                        foreach (var item in ribbonList.Items)
+                            UpdateImageThemes(item as RibbonItem, isLight);
+                        break;
                     case RibbonItem ribbonButton:
                         UpdateRibbonItemTheme(ribbonButton);
                         break;